Convert from JS to TS
